# contracts/openapi/account-service.yaml
openapi: 3.1.0
info:
  title: LedgerFlow Account Service API
  version: 1.0.0
  description: >-
    Creates and reads single-currency accounts and their immutable ledger history.
    The test-funding operation is synthetic and is exposed only when the local or
    test Spring profile is active; it is not a production banking API.
servers:
  - url: http://localhost:8081
    description: Local Account Service
tags:
  - name: Accounts
  - name: Synthetic funding
    description: Local/test-only development capability
paths:
  /api/v1/accounts:
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create an active account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
            example:
              ownerReference: customer-001
              currency: EUR
      responses:
        '201':
          description: Account created
          headers:
            Location:
              description: Relative URI of the created account
              schema:
                type: string
              example: /api/v1/accounts/3c1abf95-f4ab-4e3a-b0c8-320773339dd8
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnsupportedCurrency'
  /api/v1/accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Retrieve current account state
      responses:
        '200':
          description: Current account state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/accounts/{accountId}/ledger:
    parameters:
      - $ref: '#/components/parameters/AccountId'
      - name: page
        in: query
        required: false
        description: Zero-based page number
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        required: false
        description: Number of entries per page, bounded to protect the service
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
    get:
      tags: [Accounts]
      operationId: listLedgerEntries
      summary: List immutable ledger entries newest first
      responses:
        '200':
          description: Requested ledger page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/accounts/{accountId}/test-funding:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    post:
      tags: [Synthetic funding]
      operationId: addSyntheticTestFunding
      summary: Add a synthetic credit in local or test profiles
      description: >-
        Creates an immutable credit entry and updates the available balance in one
        PostgreSQL transaction. This route does not exist outside the local and test
        profiles. It does not represent a payment-network integration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestFundingRequest'
            example:
              amount: '1000.00'
              reference: initial-funding-001
      responses:
        '201':
          description: Synthetic funding accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Funding reference already exists for this account, or the account cannot be mutated
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/UnsupportedCurrency'
components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: The request body, path, query parameters, or monetary value is invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Request validation failed
            status: 400
            detail: One or more request fields are invalid
            instance: /api/v1/accounts
            errors:
              - field: ownerReference
                message: must not be blank
    NotFound:
      description: The account does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Account not found
            status: 404
            detail: Account 00000000-0000-0000-0000-000000000000 does not exist
            instance: /api/v1/accounts/00000000-0000-0000-0000-000000000000
    UnsupportedCurrency:
      description: Currency is syntactically valid but unsupported or mismatched
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Unsupported currency
            status: 422
            detail: Currency CHF is not supported
            instance: /api/v1/accounts
  schemas:
    CreateAccountRequest:
      type: object
      required: [ownerReference, currency]
      properties:
        ownerReference:
          type: string
          minLength: 1
          maxLength: 100
          description: Stable external owner reference; whitespace-only values are rejected
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: One supported ISO 4217 code; currently EUR, USD, or GBP
          examples: [EUR]
    TestFundingRequest:
      type: object
      required: [amount, reference]
      properties:
        amount:
          type: string
          pattern: '^(0|[1-9][0-9]{0,16})(\.[0-9]{1,2})?$'
          description: Positive decimal amount with at most two fractional digits; zero is rejected
          examples: ['1000.00']
        reference:
          type: string
          minLength: 1
          maxLength: 100
          description: Idempotency reference unique within the account
          examples: [initial-funding-001]
    Account:
      type: object
      required:
        - accountId
        - ownerReference
        - currency
        - status
        - availableBalance
        - reservedBalance
        - version
        - createdAt
        - updatedAt
      properties:
        accountId:
          type: string
          format: uuid
        ownerReference:
          type: string
        currency:
          type: string
          enum: [EUR, USD, GBP]
        status:
          type: string
          enum: [ACTIVE, FROZEN, CLOSED]
        availableBalance:
          $ref: '#/components/schemas/MoneyString'
        reservedBalance:
          $ref: '#/components/schemas/MoneyString'
        version:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        accountId: 3c1abf95-f4ab-4e3a-b0c8-320773339dd8
        ownerReference: customer-001
        currency: EUR
        status: ACTIVE
        availableBalance: '0.00'
        reservedBalance: '0.00'
        version: 0
        createdAt: '2026-07-18T12:00:00Z'
        updatedAt: '2026-07-18T12:00:00Z'
    LedgerEntry:
      type: object
      required: [ledgerEntryId, accountId, type, amount, currency, reference, createdAt]
      properties:
        ledgerEntryId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          type: string
          enum: [CREDIT, DEBIT]
        amount:
          $ref: '#/components/schemas/MoneyString'
        currency:
          type: string
          enum: [EUR, USD, GBP]
        reference:
          type: string
        createdAt:
          type: string
          format: date-time
    LedgerPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntry'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    FundingResult:
      type: object
      required: [account, ledgerEntry]
      properties:
        account:
          $ref: '#/components/schemas/Account'
        ledgerEntry:
          $ref: '#/components/schemas/LedgerEntry'
    MoneyString:
      type: string
      pattern: '^[0-9]{1,17}\.[0-9]{2}$'
      description: Non-negative NUMERIC(19,2) rendered as a JSON string
      examples: ['1000.00']
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string